# docs/openapi-frontend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MeController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/guilds": {
      "get": {
        "operationId": "GuildsController_getGuilds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/guilds/{guildId}/analytics": {
      "get": {
        "operationId": "AnalyticsController_getCombinedAnalytics",
        "parameters": [
          {
            "name": "x-timezone",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/analytics/overview": {
      "get": {
        "operationId": "AnalyticsController_getOverview",
        "parameters": [
          {
            "name": "x-timezone",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/analytics/activity-chart": {
      "get": {
        "operationId": "AnalyticsController_getActivityChart",
        "parameters": [
          {
            "name": "x-timezone",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/analytics/top-members": {
      "get": {
        "operationId": "AnalyticsController_getTopMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/heatmap": {
      "get": {
        "operationId": "HeatmapController_getHeatmap",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/leaderboard": {
      "get": {
        "operationId": "HeatmapController_getLeaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Frontend API",
    "description": "Public API. Authentication: Bearer JWT in Authorization header. All requests/responses use application/json. Envelope pattern: success in data, errors in error.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
...
      "PatchMeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}